Версия 1.2.0 Составные атрибуты и некоторые другие вещи поправлена библиотека libfloader
git-svn-id: svn+ssh://svn.code.sf.net/p/dynamicdocs/code@551 05742d63-fc59-4ccd-b3f0-d7e2a2a8a88d

+++ ddocserver/internals/functions/general/eiogetindicators.sql
@@ -7,7 +7,7 @@ create type h_eio_get_indicators as(
                                   attr_code varchar,  --при информационном обмене (в ф-и ioGetObjectAttrsEx() в качестве значения данного поля используется unique_id)
                                   id_attr_type int4,
                                   id_attr_category int4,
-                                  id_attr_value int4, 
+                                  id_attr_value int8, 
                                   start_time timestamp,
                                   stop_time timestamp,
                                   insert_time timestamp,
